requirements.txt
aiohttp==3.5.4
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
cleverbotfree==1.2.1
discord.py==1.2.5
Django==3.0.1
idna==2.8
multidict==4.7.2
pytz==2019.3
selenium==3.141.0
South==1.0.2
sqlparse==0.3.0
urllib3==1.25.7
Voice==0.1.0
websockets==6.0
yarl==1.4.2
